Fix delete button in cart

+++ react/src/Cart.jsx
@@ -126,6 +126,7 @@ function Cart() {
   const navigate = useNavigate();
   const [showConfirmation, setShowConfirmation] = useState(false);
   const [customHighTotal, setCustomHighTotal] = useState(0);
+  const [deleteProduct, setDeleteProduct] = useState("");
 
   const checkout = (order) => {
     if (order['total_cost'] > 0) {
@@ -330,6 +331,12 @@ function Cart() {
     })
   }, []);
 
+
+  const confirmDelete = (product) => {
+    setDeleteProduct(product);
+    setShowConfirmation(true);
+  }
+  
   return (
     <div className="mycart">
       {userId ?
@@ -409,23 +416,23 @@ function Cart() {
                     </h2>
                     <br /><br />
                     <h2>
-                      <button onClick={() => setShowConfirmation(true)} className="CartRemoveProductButton">
+                      <button onClick={() => confirmDelete(product)} className="CartRemoveProductButton">
                         Delete
                       </button>
                     </h2>
-                    {showConfirmation && (
-                      <div className="confirmation-modal">
-                        <div className="confirmation-dialog">
-                          <h3>Confirm Delete</h3>
-                          <p>Are you sure you want to remove "{product.product_name}" from your cart?</p>
-                          <div className="confirmation-buttons">
-                            <button onClick={() => setShowConfirmation(false)}>Cancel</button>
-                            <button onClick={() => deleteFromCart(product, order)} className="delete-button">Delete</button>
-                          </div>
+                  </div>
+                  {showConfirmation && deleteProduct === product &&
+                    <div className="confirmation-modal">
+                      <div className="confirmation-dialog">
+                        <h3>Confirm Delete</h3>
+                        <p>Are you sure you want to remove "{product.product_name}" from your cart?</p>
+                        <div className="confirmation-buttons">
+                          <button onClick={() => setShowConfirmation(false)}>Cancel</button>
+                          <button onClick={() => deleteFromCart(product, order)} className="delete-button">Delete</button>
                         </div>
                       </div>
-                    )}
-                  </div>
+                    </div>
+                  }
                   <div className="productsCell">
                     <br /><br /><br /><br /><br /><br />
                     <h2>${setPrice(product.price, product.product_type, product.size) * product.product_quantity}</h2>
@@ -465,23 +472,23 @@ function Cart() {
                     </h2>
                     <br /><br />
                     <h2>
-                      <button onClick={() => setShowConfirmation(true)} className="CartRemoveProductButton">
+                      <button onClick={() => confirmDelete(product)} className="CartRemoveProductButton">
                         Delete
                       </button>
                     </h2>
-                    {showConfirmation && (
-                      <div className="confirmation-modal"> {console.log(product)}
-                        <div className="confirmation-dialog">
-                          <h3>Confirm Delete</h3>
-                          <p>Are you sure you want to remove "{product.product_name}" from your cart?</p>
-                          <div className="confirmation-buttons">
-                            <button onClick={() => setShowConfirmation(false)}>Cancel</button>
-                            <button onClick={() => deleteFromCart(product, order)} className="delete-button">Delete</button>
-                          </div>
+                  </div>
+                  {showConfirmation && deleteProduct === product &&
+                    <div className="confirmation-modal">
+                      <div className="confirmation-dialog">
+                        <h3>Confirm Delete</h3>
+                        <p>Are you sure you want to remove "{product.product_name}" from your cart?</p>
+                        <div className="confirmation-buttons">
+                          <button onClick={() => setShowConfirmation(false)}>Cancel</button>
+                          <button onClick={() => deleteFromCart(product, order)} className="delete-button">Delete</button>
                         </div>
                       </div>
-                    )}
-                  </div>
+                    </div>
+                  }
                   <div className="productsCell">
                     <br /><br /><br /><br /><br /><br />
                     <h2>${setPrice(product.price, product.product_type, product.size) * product.product_quantity} - ${(setPrice(product.price, product.product_type, product.size) + 6) * product.product_quantity}</h2>
